update: separate sketch logic to component
